refactor(pwa): Lot 3 frontend refactoring (#89)
- Extract shared constants, weather icons, and URL validation helpers
- Replace raw fetch calls with downloadStageFile() and scrapeAccommodation()
- Remove all `as Parameters<typeof t>` type assertion hacks
- Extract useTripActions hook from trip-planner (363 → ~100 lines)
- Extract StageDownloads and StageDistanceEditor sub-components
- Add Next.js App Router error boundary
- Replace `as` casts in parseApiError with type guard functions
