+++ pwa/src/components/trip-summary.tsx
@@ -1,41 +1,8 @@
 import { useTranslations } from "next-intl";
-import {
-  Sun,
-  CloudSun,
-  Cloud,
-  CloudRain,
-  CloudLightning,
-  Snowflake,
-  CloudFog,
-  ArrowUp,
-  ArrowDown,
-  Bike,
-  Mountain,
-  Info,
-} from "lucide-react";
+import { ArrowUp, ArrowDown, Bike, Mountain, Info } from "lucide-react";
 import { Skeleton } from "@/components/ui/skeleton";
+import { weatherIconMap, DefaultWeatherIcon } from "@/lib/weather-icons";
 import type { WeatherData } from "@/lib/validation/schemas";
-
-const weatherIconMap: Record<string, React.ElementType> = {
-  "01d": Sun,
-  "01n": Sun,
-  "02d": CloudSun,
-  "02n": CloudSun,
-  "03d": Cloud,
-  "03n": Cloud,
-  "04d": Cloud,
-  "04n": Cloud,
-  "09d": CloudRain,
-  "09n": CloudRain,
-  "10d": CloudRain,
-  "10n": CloudRain,
-  "11d": CloudLightning,
-  "11n": CloudLightning,
-  "13d": Snowflake,
-  "13n": Snowflake,
-  "50d": CloudFog,
-  "50n": CloudFog,
-};
 
 interface TripSummaryProps {
   totalDistance: number | null;
@@ -61,7 +28,9 @@ export function TripSummary({
   if (!showSkeleton && totalDistance === null && totalElevation === null)
     return null;
 
-  const WeatherIcon = weather ? (weatherIconMap[weather.icon] ?? Cloud) : Cloud;
+  const WeatherIcon = weather
+    ? (weatherIconMap[weather.icon] ?? DefaultWeatherIcon)
+    : DefaultWeatherIcon;
 
   return (
     <div className="space-y-2">
